feat: added basic layout
Resolves #6

- Added basic layout components and styles.
