+++ components/Journal.vue
@@ -1,0 +1,3 @@
+<template>
+  <Section title="Journal"> </Section>
+</template>
