+++ components/Milestones.vue
@@ -1,0 +1,3 @@
+<template>
+  <Section title="Milestones"> </Section>
+</template>
