+++ components/MoodTracker.vue
@@ -1,0 +1,3 @@
+<template>
+  <Section title="Mood tracker"> </Section>
+</template>
