+++ components/Tasks.vue
@@ -1,0 +1,3 @@
+<template>
+  <Section title="Tasks"> </Section>
+</template>
